+++ addons/auth_totp/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import controllers
 from . import models
+from . import wizard
