+++ addons/auth_totp/views/res_users_views.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record model="ir.ui.view" id="res_users_view_search">
+        <field name="name">res.users.view.search.inherit.auth.totp</field>
+        <field name="model">res.users</field>
+        <field name="inherit_id" ref="base.view_users_search" />
+        <field name="arch" type="xml">
+            <xpath expr="//search" position="inside">
+                <separator/>
+                <filter name="totp_enabled" string="Two-factor authentication Enabled" domain="[('totp_secret','!=',False)]"/>
+                <separator/>
+                <filter name="totp_disabled" string="Two-factor authentication Disabled" domain="[('totp_secret','=',False)]"/>
+            </xpath>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_totp_form">
+        <field name="name">user form: add totp status</field>
+        <field name="model">res.users</field>
+        <field name="inherit_id" ref="base.view_users_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//page[@name='preferences']" position="after">
+                <page string="Account Security" name="security" attrs="{'invisible': [('id', '=', False)]}">
+                    <field name="totp_enabled" invisible="1"/>
+                    <!-- For own user, allow to activate the two-factor Authentication -->
+                    <group>
+                        <div>
+                            <div class="o_horizontal_separator d-flex align-items-center mt-0">Two-factor Authentication
+                                <div attrs="{'invisible': [('totp_enabled', '!=', False)]}">
+                                    <button attrs="{'invisible': &quot;[('id', '=', uid)]&quot;}" name="action_totp_enable_wizard"
+                                            disabled="1" type="object" class="fa fa-toggle-off o_auth_2fa_btn disabled" aria-label="Enable 2FA"></button>
+                                    <button attrs="{'invisible': &quot;[('id', '!=', uid)]&quot;}" name="action_totp_enable_wizard"
+                                        type="object" class="fa fa-toggle-off o_auth_2fa_btn disabled" aria-label="Enable 2FA"></button>
+                                    <button groups="base.group_erp_manager" attrs="{'invisible': &quot;[('id', '=', uid)]&quot;}"
+                                        name="action_totp_invite" string="Invite to use 2FA" type="object" class="btn btn-secondary"></button>
+                                </div>
+                                <button attrs="{'invisible': [('totp_enabled', '=', False)]}" name="action_totp_disable" type="object"
+                                        class="fa fa-toggle-on o_auth_2fa_btn text-primary enabled" aria-label="Disable 2FA"></button>
+                            </div>
+                            <span attrs="{'invisible': [('totp_enabled', '!=', False)]}" class="text-muted">
+                                Two-factor Authentication ("2FA") is a system of double authentication.
+                                The first one is done with your password and the second one with a code you get from a dedicated mobile app.
+                                Popular ones include Authy, Google Authenticator or the Microsoft Authenticator.
+                                <a href="https://www.odoo.com/documentation/15.0/applications/general/auth/2fa.html"
+                                   title="Learn More" target="_blank">Learn More</a>
+                            </span>
+                            <span attrs="{'invisible': [('totp_enabled', '=', False)]}" class="text-muted">Your account is protected!</span>
+                        </div>
+                    </group>
+                </page>
+            </xpath>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_totp_field">
+        <field name="name">users preference: totp</field>
+        <field name="model">res.users</field>
+        <field name="inherit_id" ref="base.view_users_form_simple_modif"/>
+        <field name="arch" type="xml">
+            <group name="auth" position="after">
+                <field name="totp_enabled" invisible="1"/>
+                <group>
+                    <div>
+                        <div class="o_horizontal_separator mt-0">Two-factor Authentication
+                            <button attrs="{'invisible': [('totp_enabled', '!=', False)]}" name="action_totp_enable_wizard"
+                               type="object" class="fa fa-toggle-off o_auth_2fa_btn mb-1" aria-label="Enable 2FA"/>
+                            <button attrs="{'invisible': [('totp_enabled', '=', False)]}" name="action_totp_disable"
+                               type="object" class="fa fa-toggle-on o_auth_2fa_btn text-primary" aria-label="Disable 2FA"/>
+                        </div>
+                        <span attrs="{'invisible': [('totp_enabled', '!=', False)]}" class="text-muted">
+                            Two-factor Authentication ("2FA") is a system of double authentication.
+                            The first one is done with your password and the second one with a code you get from a dedicated mobile app.
+                            Popular ones include Authy, Google Authenticator or the Microsoft Authenticator.
+                            <a href="https://www.odoo.com/documentation/15.0/applications/general/auth/2fa.html"
+                               title="Learn More" target="_blank">Learn More</a>
+                        </span>
+                        <span attrs="{'invisible': [('totp_enabled', '=', False)]}" class="text-muted">Your account is protected!</span>
+                    </div>
+                </group>
+            </group>
+        </field>
+    </record>
+
+    <!-- View used when coming from "invite to use 2FA" mail -->
+    <record model="ir.ui.view" id="res_users_view_form_security">
+        <field name="name">users preference: Account Security</field>
+        <field name="model">res.users</field>
+        <field name="inherit_id" ref="base.view_users_form_simple_modif"/>
+        <field name="mode">primary</field>
+        <field name="arch" type="xml">
+            <form position="attributes">
+                <attribute name='create'>0</attribute>
+                <attribute name='edit'>0</attribute>
+                <attribute name='delete'>0</attribute>
+            </form>
+            <h1 position="replace"/>
+            <xpath expr="//field[@name='image_1920']" position="replace"/>
+            <notebook position="replace">
+                <header>
+                </header>
+                <sheet>$0</sheet>
+            </notebook>
+            <notebook position="before">
+                <field name="image_1920" widget="image" class="oe_avatar" options="{'zoom': true, 'preview_image':'image_128'}"/>
+                    <div class="oe_title">
+                        <h1>
+                            <field name="name" placeholder="Name" required="True" readonly="context.get('from_my_profile', False)"/>
+                        </h1>
+                    </div>
+            </notebook>
+            <page name="preferences_page" position="replace"></page>
+            <footer position="replace"/>
+        </field>
+    </record>
+</odoo>
