+++ addons/auth_totp/wizard/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import ir_http
-from . import res_users
-from . import totp
+from . import auth_totp_wizard
